Destroy/Add model
Instead of bootstrapping with a default model we just remove the default model
and add a new model to the environment with our custom name. In the case where
'controller' is the suggested-model we do nothing and allow the deployment to
happen on that controller model. For example, landscape uses this to not utilize
a second machine just for the landscape bits and instead deploy everything on
the bootstrap(controller model, effectively machine 0)

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>